+++ src/webparts/recruitmentApp/Models/RecuritmentVRR.ts
@@ -388,7 +388,7 @@ export type ValidationErrors = {
 };
 
 export type tabCount = {
-  AssignHRCount: number;
+  AssignHRCount?: number;
   UploadONEMCount: number;
   UploadAdvertisementCount: number;
   AssignAgencyCount: number;
@@ -396,7 +396,7 @@ export type tabCount = {
   ReviewHODCount: number;
   lineManagerInterviewCount: number;
   HODReviewScoreCount: number;
-  EvaluationCount: number;
+  EvaluationCount?: number;
   advertExtensionCount: number;
 }
