+++ src/webparts/recruitmentApp/Screens/ReviewScoreCard/CandidateList.tsx
